lots of work on implementation, partially working

+++ src/misc/Point.java
@@ -1,20 +1,20 @@
 package misc;
 
 public class Point {
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 	
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 }
